# api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: AI Store Manage - 用户与部门管理 API
  description: |
    WMS 仓储管理系统 — 用户与部门模块 API 契约。

    本契约基于 `sys_user` 表字段设计，遵循 Contract First 原则，
    定义用户 CRUD、状态变更、密码重置等接口。

    **设计原则：**
    - RESTful 风格，资源路径统一为 `/api/users`
    - 严格区分数据库字段与 API 暴露字段
    - `password` 仅出现在写入类请求，永不出现在响应中
    - `deleted` 为内部逻辑删除标记，不暴露给业务接口
    - 枚举字段（gender、status）明确约束取值范围
    - 统一错误响应结构，便于前端统一处理
  version: 1.0.0
  contact:
    name: AI Store Manage Team
servers:
  - url: http://localhost:8080
    description: 本地开发环境
...
paths:
...
  /api/locations:
    post:
      tags: [库位管理]
      summary: 创建库位
      operationId: createLocation
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateLocationRequest' } } }
      responses:
        '201': { description: 创建成功, content: { application/json: { schema: { $ref: '#/components/schemas/Location' } } } }
        '400': { description: 校验失败, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: 仓库/库区不存在, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '409': { description: 编码冲突, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    get:
      tags: [库位管理]
      summary: 查询库位列表
      operationId: listLocations
      parameters:
        - { name: keyword, in: query, required: false, schema: { type: string, maxLength: 64 } }
        - { name: warehouseId, in: query, required: false, schema: { type: integer, format: int64 } }
        - { name: zoneId, in: query, required: false, schema: { type: integer, format: int64 } }
        - { name: status, in: query, required: false, schema: { type: integer, enum: [0, 1] } }
        - { name: page, in: query, required: false, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: pageSize, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200': { description: 查询成功, content: { application/json: { schema: { $ref: '#/components/schemas/LocationListResponse' } } } }
  /api/locations/{id}:
    get:
      tags: [库位管理]
      summary: 查询库位详情
      operationId: getLocationById
      parameters: [ { $ref: '#/components/parameters/LocationId' } ]
      responses:
        '200': { description: 查询成功, content: { application/json: { schema: { $ref: '#/components/schemas/Location' } } } }
        '404': { description: 不存在, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    put:
      tags: [库位管理]
      summary: 更新库位
      operationId: updateLocation
      parameters: [ { $ref: '#/components/parameters/LocationId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UpdateLocationRequest' } } }
      responses:
        '200': { description: 更新成功, content: { application/json: { schema: { $ref: '#/components/schemas/Location' } } } }
        '400': { description: 校验失败, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: 不存在, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    delete:
      tags: [库位管理]
      summary: 删除库位
      operationId: deleteLocation
      parameters: [ { $ref: '#/components/parameters/LocationId' } ]
      responses:
        '204': { description: 删除成功 }
        '404': { description: 不存在, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /api/locations/{id}/status:
    patch:
      tags: [库位管理]
      summary: 变更库位状态
      operationId: updateLocationStatus
      parameters: [ { $ref: '#/components/parameters/LocationId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UpdateStatusRequest' } } }
      responses:
        '200': { description: 变更成功, content: { application/json: { schema: { $ref: '#/components/schemas/Location' } } } }
        '404': { description: 不存在, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

  # ── 托盘实例管理 ──
  /api/lpns:
    post:
      tags: [托盘实例管理]
      summary: 创建托盘实例
      operationId: createLpn
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateLpnRequest' } } }
      responses:
        '201': { description: 创建成功, content: { application/json: { schema: { $ref: '#/components/schemas/Lpn' } } } }
        '400': { description: 校验失败, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: 托盘类型/仓库/库位不存在, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '409': { description: 托盘号冲突, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    get:
      tags: [托盘实例管理]
      summary: 查询托盘实例列表
      operationId: listLpns
      parameters:
        - { name: keyword, in: query, required: false, schema: { type: string, maxLength: 64 } }
        - { name: warehouseId, in: query, required: false, schema: { type: integer, format: int64 } }
        - { name: status, in: query, required: false, schema: { $ref: '#/components/schemas/LpnStatus' } }
        - { name: page, in: query, required: false, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: pageSize, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200': { description: 查询成功, content: { application/json: { schema: { $ref: '#/components/schemas/LpnListResponse' } } } }
  /api/lpns/{id}:
    get:
      tags: [托盘实例管理]
      summary: 查询托盘实例详情
      operationId: getLpnById
      parameters: [ { $ref: '#/components/parameters/LpnId' } ]
      responses:
        '200': { description: 查询成功, content: { application/json: { schema: { $ref: '#/components/schemas/Lpn' } } } }
        '404': { description: 不存在, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    put:
      tags: [托盘实例管理]
      summary: 更新托盘实例
      operationId: updateLpn
      parameters: [ { $ref: '#/components/parameters/LpnId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UpdateLpnRequest' } } }
      responses:
        '200': { description: 更新成功, content: { application/json: { schema: { $ref: '#/components/schemas/Lpn' } } } }
        '400': { description: 校验失败, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: 不存在, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    delete:
      tags: [托盘实例管理]
      summary: 删除托盘实例
      operationId: deleteLpn
      parameters: [ { $ref: '#/components/parameters/LpnId' } ]
      responses:
        '204': { description: 删除成功 }
        '404': { description: 不存在, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /api/lpns/{id}/status:
    patch:
      tags: [托盘实例管理]
      summary: 变更托盘实例状态（在库/在途/空置）
      operationId: updateLpnStatus
      parameters: [ { $ref: '#/components/parameters/LpnId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UpdateLpnStatusRequest' } } }
      responses:
        '200': { description: 变更成功, content: { application/json: { schema: { $ref: '#/components/schemas/Lpn' } } } }
        '404': { description: 不存在, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

  # ── 库存管理 ──
  /api/inventory:
    post:
      tags: [库存管理]
      summary: 创建库存记录
      operationId: createInventory
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateInventoryRequest' } } }
      responses:
        '201': { description: 创建成功, content: { application/json: { schema: { $ref: '#/components/schemas/Inventory' } } } }
        '400': { description: 校验失败, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: SKU 不存在, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    get:
      tags: [库存管理]
      summary: 查询库存列表
      operationId: listInventory
      parameters:
        - { name: skuId, in: query, required: false, schema: { type: integer, format: int64 } }
        - { name: lpnId, in: query, required: false, schema: { type: integer, format: int64 } }
        - { name: locationId, in: query, required: false, schema: { type: integer, format: int64 } }
        - { name: page, in: query, required: false, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: pageSize, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200': { description: 查询成功, content: { application/json: { schema: { $ref: '#/components/schemas/InventoryListResponse' } } } }
  /api/inventory/summary:
    get:
      tags: [库存管理]
      summary: 库存统计（库存数量 + 托盘数量 + 整托/尾托）
      operationId: getInventorySummary
      parameters:
        - { name: skuId, in: query, required: true, schema: { type: integer, format: int64 } }
        - { name: warehouseId, in: query, required: false, schema: { type: integer, format: int64 } }
      responses:
        '200': { description: 统计成功, content: { application/json: { schema: { $ref: '#/components/schemas/InventorySummary' } } } }
        '404': { description: SKU 不存在, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /api/inventory/{id}:
    get:
      tags: [库存管理]
      summary: 查询库存详情
      operationId: getInventoryById
      parameters: [ { $ref: '#/components/parameters/InventoryId' } ]
      responses:
        '200': { description: 查询成功, content: { application/json: { schema: { $ref: '#/components/schemas/Inventory' } } } }
        '404': { description: 不存在, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    put:
      tags: [库存管理]
      summary: 更新库存记录
      operationId: updateInventory
      parameters: [ { $ref: '#/components/parameters/InventoryId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/UpdateInventoryRequest' } } }
      responses:
        '200': { description: 更新成功, content: { application/json: { schema: { $ref: '#/components/schemas/Inventory' } } } }
        '400': { description: 校验失败, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '404': { description: 不存在, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    delete:
      tags: [库存管理]
      summary: 删除库存记录
      operationId: deleteInventory
      parameters: [ { $ref: '#/components/parameters/InventoryId' } ]
      responses:
        '204': { description: 删除成功 }
        '404': { description: 不存在, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }

  # ── 车辆管理 ──
...
components:
  parameters:
...
    LocationId:
      name: id
      in: path
      description: 库位 ID
      required: true
      schema: { type: integer, format: int64, example: 1 }
    LpnId:
      name: id
      in: path
      description: 托盘实例 ID
      required: true
      schema: { type: integer, format: int64, example: 1 }
    InventoryId:
      name: id
      in: path
      description: 库存记录 ID
      required: true
      schema: { type: integer, format: int64, example: 1 }
...
  schemas:
...
    UpdateStatusRequest:
      type: object
      description: 通用状态变更请求体（0=禁用,1=启用）
      required: [status]
      properties:
        status: { type: integer, enum: [0, 1], example: 0 }
...
    LpnStatus:
      type: string
      enum: [IN_STOCK, IN_TRANSIT, EMPTY]
      description: 托盘状态 在库/在途/空置
    UpdateLpnStatusRequest:
      type: object
      required: [status]
      properties:
        status: { $ref: '#/components/schemas/LpnStatus' }

    # 库位
    CreateLocationRequest:
      type: object
      required: [warehouseId, code]
      properties:
        warehouseId: { type: integer, format: int64 }
        zoneId: { type: integer, format: int64, nullable: true, description: 所属库区（可空） }
        code: { type: string, minLength: 1, maxLength: 32, pattern: '^[A-Za-z0-9\-]+$', example: "A-01-01" }
        locType: { type: string, maxLength: 32, description: "货架/地堆" }
        status: { type: integer, enum: [0, 1], default: 1 }
    UpdateLocationRequest:
      type: object
      properties:
        code: { type: string, minLength: 1, maxLength: 32, pattern: '^[A-Za-z0-9\-]+$', description: 库位编码（仓库内唯一，可改） }
        zoneId: { type: integer, format: int64 }
        locType: { type: string, maxLength: 32 }
    Location:
      type: object
      properties:
        id: { type: integer, format: int64 }
        warehouseId: { type: integer, format: int64 }
        warehouseName: { type: string, nullable: true }
        zoneId: { type: integer, format: int64, nullable: true }
        zoneName: { type: string, nullable: true }
        code: { type: string }
        locType: { type: string, nullable: true }
        status: { type: integer, enum: [0, 1] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        createdBy: { type: integer, format: int64, nullable: true }
        updatedBy: { type: integer, format: int64, nullable: true }
    LocationSummary:
      type: object
      properties:
        id: { type: integer, format: int64 }
        warehouseId: { type: integer, format: int64 }
        warehouseName: { type: string, nullable: true }
        zoneName: { type: string, nullable: true }
        code: { type: string }
        locType: { type: string, nullable: true }
        status: { type: integer, enum: [0, 1] }
        createdAt: { type: string, format: date-time }
    LocationListResponse:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/LocationSummary' } }
        total: { type: integer, format: int64 }
        page: { type: integer }
        pageSize: { type: integer }
        totalPages: { type: integer }

    # 托盘实例 LPN
    CreateLpnRequest:
      type: object
      required: [lpnCode, palletTypeId, warehouseId]
      properties:
        lpnCode: { type: string, minLength: 1, maxLength: 64, description: SSCC 托盘号 }
        palletTypeId: { type: integer, format: int64 }
        warehouseId: { type: integer, format: int64 }
        locationId: { type: integer, format: int64, nullable: true }
        status: { allOf: [ { $ref: '#/components/schemas/LpnStatus' } ], default: IN_STOCK }
        grossWeight: { type: number, nullable: true }
    UpdateLpnRequest:
      type: object
      properties:
        locationId: { type: integer, format: int64 }
        grossWeight: { type: number }
    Lpn:
      type: object
      properties:
        id: { type: integer, format: int64 }
        lpnCode: { type: string }
        palletTypeId: { type: integer, format: int64 }
        palletTypeName: { type: string, nullable: true }
        warehouseId: { type: integer, format: int64 }
        warehouseName: { type: string, nullable: true }
        locationId: { type: integer, format: int64, nullable: true }
        locationCode: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/LpnStatus' }
        grossWeight: { type: number, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        createdBy: { type: integer, format: int64, nullable: true }
        updatedBy: { type: integer, format: int64, nullable: true }
    LpnSummary:
      type: object
      properties:
        id: { type: integer, format: int64 }
        lpnCode: { type: string }
        palletTypeName: { type: string, nullable: true }
        warehouseName: { type: string, nullable: true }
        locationCode: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/LpnStatus' }
        createdAt: { type: string, format: date-time }
    LpnListResponse:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/LpnSummary' } }
        total: { type: integer, format: int64 }
        page: { type: integer }
        pageSize: { type: integer }
        totalPages: { type: integer }

    # 库存
    CreateInventoryRequest:
      type: object
      required: [skuId, qtyOnHand]
      properties:
        skuId: { type: integer, format: int64 }
        lpnId: { type: integer, format: int64, nullable: true, description: 所在托盘（可空，地面库存） }
        locationId: { type: integer, format: int64, nullable: true }
        lotNo: { type: string, maxLength: 64, nullable: true }
        mfgDate: { type: string, format: date, nullable: true }
        expDate: { type: string, format: date, nullable: true }
        qtyOnHand: { type: number, description: 在库（基本单位） }
        qtyReserved: { type: number, default: 0, description: 锁定 }
    UpdateInventoryRequest:
      type: object
      properties:
        lpnId: { type: integer, format: int64 }
        locationId: { type: integer, format: int64 }
        lotNo: { type: string, maxLength: 64 }
        mfgDate: { type: string, format: date }
        expDate: { type: string, format: date }
        qtyOnHand: { type: number }
        qtyReserved: { type: number }
    Inventory:
      type: object
      properties:
        id: { type: integer, format: int64 }
        skuId: { type: integer, format: int64 }
        skuCode: { type: string, nullable: true }
        skuName: { type: string, nullable: true }
        lpnId: { type: integer, format: int64, nullable: true }
        lpnCode: { type: string, nullable: true }
        locationId: { type: integer, format: int64, nullable: true }
        locationCode: { type: string, nullable: true }
        lotNo: { type: string, nullable: true }
        mfgDate: { type: string, format: date, nullable: true }
        expDate: { type: string, format: date, nullable: true }
        qtyOnHand: { type: number }
        qtyReserved: { type: number }
        qtyAvailable: { type: number, description: qtyOnHand - qtyReserved（计算） }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        createdBy: { type: integer, format: int64, nullable: true }
        updatedBy: { type: integer, format: int64, nullable: true }
    InventorySummaryItem:
      type: object
      properties:
        id: { type: integer, format: int64 }
        skuId: { type: integer, format: int64 }
        skuName: { type: string, nullable: true }
        lpnCode: { type: string, nullable: true, description: 托盘名称（SSCC 托盘号） }
        palletTypeName: { type: string, nullable: true, description: 托盘类型名（经 LPN 关联） }
        locationCode: { type: string, nullable: true }
        lotNo: { type: string, nullable: true }
        qtyOnHand: { type: number }
        qtyReserved: { type: number }
        qtyAvailable: { type: number }
        createdAt: { type: string, format: date-time }
    InventoryListResponse:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/InventorySummaryItem' } }
        total: { type: integer, format: int64 }
        page: { type: integer }
        pageSize: { type: integer }
        totalPages: { type: integer }

    # 库存统计（需求②：库存数量 + 托盘数量 + 整托/尾托）
    InventoryPallet:
      type: object
      properties:
        lpnId: { type: integer, format: int64 }
        lpnCode: { type: string, nullable: true }
        qty: { type: number, description: 该托盘上该 SKU 的在库数量（基本单位） }
        fullPallet: { type: boolean, nullable: true, description: 是否整托（对比标准每托数；无标准则为 null） }
    InventorySummary:
      type: object
      properties:
        skuId: { type: integer, format: int64 }
        skuName: { type: string, nullable: true }
        totalQty: { type: number, description: 总库存 SUM(qty_on_hand) }
        totalReserved: { type: number, description: 总锁定 }
        totalAvailable: { type: number, description: 总可用 }
        palletCount: { type: integer, description: 托盘数量 COUNT(DISTINCT lpn_id) }
        recordCount: { type: integer, description: 库存记录数 }
        standardPalletQty: { type: number, nullable: true, description: 标准每托数（来自 packaging_relation 定量整托；无则 null） }
        pallets: { type: array, items: { $ref: '#/components/schemas/InventoryPallet' }, description: 每托明细（整托/尾托） }

    # ──────────────────────────────────────────────
    # 运输管理 Schema（车辆 / 打油 / 打卡 / 文件）
    # ──────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      description: 统一错误响应体
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: 错误码（大写下划线风格，用于前端判断错误类型）
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: 错误描述（人类可读）
          example: "请求参数校验失败"
        details:
          type: array
          description: 字段级错误详情（参数校验失败时返回）
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      description: 字段级错误详情
      properties:
        field:
          type: string
          description: 出错字段名
          example: "username"
        message:
          type: string
          description: 字段错误描述
          example: "用户名长度必须在4-64个字符之间"